UdemyMl/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
attrs=20.2.0=py_0
beautifulsoup4=4.9.1=py37_0
blas=1.0=mkl
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
click=7.1.2=py_0
cryptography=3.1.1=py37h7a1dbc1_0
cycler=0.10.0=py37_0
flask=1.1.1=py_1
freetype=2.10.2=hd328e21_0
gunicorn=20.0.4=pypi_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=py_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
itsdangerous=1.1.0=py37_0
jinja2=2.10.1=py37_0
joblib=0.17.0=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py37_1
kiwisolver=1.2.0=py37h74a9793_0
libiconv=1.15=h1df5818_7
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_1
libxml2=2.9.10=h464c3ec_1
libxslt=1.1.34=he774522_0
lxml=4.5.1=py37h1350720_0
lz4-c=1.9.2=hf4a77e7_3
markupsafe=1.1.1=py37hfa6e2cd_1
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py37hba9282a_0
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
nltk=3.5=py_0
numpy=1.19.1=py37h5510c5b_0
numpy-base=1.19.1=py37ha3acd2a_0
olefile=0.46=py37_0
openssl=1.1.1h=he774522_0
pandas=1.1.3=py37ha925a31_0
pickleshare=0.7.5=py37_1001
pillow=7.2.0=py37hcc1f983_0
pip=20.2.3=py37_0
pycparser=2.20=py_2
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h6538335_2
pyrsistent=0.17.3=py37he774522_0
pysocks=1.7.1=py37_1
python=3.7.9=h60c2a47_0
python-dateutil=2.8.1=py_0
pytz=2020.1=py_0
qt=5.9.7=vc14h73c81de_0
regex=2020.9.27=py37he774522_0
requests=2.23.0=py37_0
scikit-learn=0.23.2=py37h47e9c7a_0
scipy=1.5.2=py37h9439919_0
seaborn=0.11.0=py_0
setuptools=50.3.0=py37h9490d1a_1
sip=4.19.8=py37h6538335_0
six=1.15.0=py_0
soupsieve=2.0.1=py_0
sqlite=3.33.0=h2a8f88b_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=he774522_0
tmdbv3api=1.6.2=pypi_0
tornado=6.0.4=py37he774522_1
tqdm=4.50.2=py_0
urllib3=1.25.9=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
werkzeug=0.15.5=py_0
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
xz=5.2.5=h62dcd97_0
zipp=3.3.0=py_0
zlib=1.2.11=h62dcd97_4
zstd=1.4.5=h04227a9_0
